fixed hub_tier attributes in country.yaml

+++ src/mirage/frameworks/epidemics/gpe.py
@@ -31,7 +31,7 @@ class Locus(EntityV1):
     infected: int
     recovered: int
     dead: int
-    area: int
+    area: float
     quarantine_facilities: int
     general_hospitals: int
     vaccine_distribution_centers: int
